+++ Nudge/UI/SimpleMode/SimpleMode.swift
@@ -32,7 +32,15 @@ struct SimpleMode: View {
             
             Text(appState.deviceSupportedByOSVersion ? getMainHeader().localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)) : getMainHeaderUnsupported().localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
                 .font(.title)
-            
+
+            if UserInterfaceVariables.simpleModeShowRequiredOSVersion {
+                requiredOSVersionView
+            }
+
+            if UserInterfaceVariables.simpleModeShowRequiredDate {
+                requiredDateView
+            }
+
             remainingTimeView
 
             if UserInterfaceVariables.showDeferralCount {
@@ -83,6 +91,22 @@ struct SimpleMode: View {
             Text(String(appState.userDeferrals))
                 .foregroundColor(infoTextColor)
                 .font(.title2)
+        }
+    }
+
+    private var requiredOSVersionView: some View {
+        HStack {
+            Text("Required OS Version:".localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
+            Text(String(appState.requiredMinimumOSVersion))
+                .foregroundColor(infoTextColor)
+        }
+    }
+
+    private var requiredDateView: some View {
+        HStack {
+            Text("Required Date:".localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
+            Text(DateManager().coerceDateToString(date: requiredInstallationDate, formatterString: UserInterfaceVariables.requiredInstallationDisplayFormat))
+                .foregroundColor(infoTextColor)
         }
     }